+++ src/main/java/domain/Card/Card.java
@@ -1,26 +1,31 @@
-package domain;
+package domain.Card;
 
 import java.util.Objects;
 
 public class Card {
-
+    
     private final CardNumber number;
-
+    
     private final CardShape shape;
-
+    
     public Card(CardNumber number, CardShape shape) {
         this.number = number;
         this.shape = shape;
     }
-
+    
     public CardNumber getCardNumber() {
         return number;
     }
-
+    
     public CardShape getCardShape() {
         return shape;
     }
-
+    
+    @Override
+    public int hashCode() {
+        return Objects.hash(number, shape);
+    }
+    
     @Override
     public boolean equals(Object o) {
         if (this == o) {
@@ -31,10 +36,5 @@ public class Card {
         }
         Card card = (Card) o;
         return number == card.number && shape == card.shape;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(number, shape);
     }
 }
