+++ src/main/java/domain/Card/CardNumber.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.Card;
 
 public enum CardNumber {
     ACE(1, "A"),
@@ -14,19 +14,19 @@ public enum CardNumber {
     JACK(10, "J"),
     QUEEN(10, "Q"),
     KING(10, "K");
-
+    
     private final int score;
     private final String name;
-
+    
     CardNumber(int score, String name) {
         this.score = score;
         this.name = name;
     }
-
+    
     public int getScore() {
         return this.score;
     }
-
+    
     public String getName() {
         return this.name;
     }
