+++ src/main/java/domain/user/Dealer.java
@@ -1,14 +1,14 @@
 package domain.user;
 
-import domain.Card;
+import domain.Card.Card;
 import java.util.List;
 
 public class Dealer extends Participant {
-
+    
     public Dealer() {
         super(DEALER_NAME);
     }
-
+    
     @Override
     public List<Card> getReadyCards() {
         if (cards.size() < 1) {
