+++ src/main/java/view/OutputView.java
@@ -1,18 +1,18 @@
 package view;
 
-import domain.Card;
-import domain.CardShape;
+import domain.Card.Card;
+import domain.Card.CardShape;
 import domain.GameResult;
 import domain.user.Participant;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class OutputView {
-
+    
     public static void printDealerReceivedCard() {
         System.out.println(System.lineSeparator() + "딜러는 16이하라 한장의 카드를 더 받았습니다.");
     }
-
+    
     public static void printDealerGameResult(GameResult dealerResult, int playerCount) {
         System.out.println(System.lineSeparator() + "## 최종 승패");
         int winCount = dealerResult.getWinCount();
@@ -31,13 +31,13 @@ public class OutputView {
         }
         System.out.println(stringBuilder);
     }
-
+    
     public static void printAllParticipantsStatus(List<Participant> allPlayers) {
         allPlayers.forEach(
                 (participant) -> OutputView.printNameCardsScore(participant.getName(), participant.getCards(),
                         participant.calculateScore()));
     }
-
+    
     public static void printNameCardsScore(String name, List<Card> cards, int score) {
         String formattedNameAndScore = name
                 + ": "
@@ -47,12 +47,12 @@ public class OutputView {
                 + score;
         System.out.println(formattedNameAndScore);
     }
-
+    
     private static String formatCard(Card card) {
         return card.getCardNumber().getName()
                 + formatCardShape(card.getCardShape());
     }
-
+    
     private static String formatCardShape(CardShape cardShape) {
         if (cardShape == CardShape.SPADE) {
             return "스페이드";
@@ -65,7 +65,7 @@ public class OutputView {
         }
         return "다이야몬드";
     }
-
+    
     public static void printPlayerResult(String name, GameResult gameResult) {
         StringBuilder stringBuilder = new StringBuilder();
         stringBuilder.append(name);
@@ -80,7 +80,7 @@ public class OutputView {
         }
         System.out.println(stringBuilder);
     }
-
+    
     public static void printReadyMessage(List<Participant> participants) {
         List<String> participantNames = participants.stream().map(Participant::getName).collect(Collectors.toList());
         StringBuilder stringBuilder = new StringBuilder();
@@ -91,12 +91,12 @@ public class OutputView {
         stringBuilder.append(System.lineSeparator());
         System.out.println(stringBuilder);
     }
-
+    
     public static void printParticipantsNameAndCards(List<Participant> allParticipant) {
         allParticipant.forEach(
                 (participant) -> OutputView.printNameAndCards(participant.getName(), participant.getReadyCards()));
     }
-
+    
     public static void printNameAndCards(String name, List<Card> cards) {
         String formattedNameAndCards = name
                 + ": "
