+++ src/test/java/domain/CardTest.java
@@ -1,11 +1,14 @@
 package domain;
 
+import domain.Card.Card;
+import domain.Card.CardNumber;
+import domain.Card.CardShape;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class CardTest {
-
+    
     @Test
     @DisplayName("카드는 자신의 점수를 반환한다.")
     void getScore() {
